glm/requirements.txt
celery==3.1.19
django==1.8.7
djangorestframework==3.3.1
django_filter==0.11.0
django-oauth-toolkit==0.9.0
gunicorn==19.4.1
markdown==2.6.5
motor==0.5
pika==0.10.0
pillow==3.0.0
psycopg2==2.6.1
pygments==2.0.2
python-dateutil==2.5.2
redis==2.10.5
scapy==2.3.1
simplejson==3.8.1
sockjs-tornado==1.0.2
tornado==4.3
tornado-redis==2.4.18
tweepy==3.5.0
yolk==0.4.3
#pcapy==0.10.9
